pyproject.toml
[tool.poetry]
name = "iocryptobot"
version = "0.1.0"
description = ""
authors = ["Adam Zugiraev <adam.zugiraev@onlinesup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.5"
python-dotenv = "^1.0.0"
ccxt = "^4.0.112"
apscheduler = "^3.10.4"
mysql-connector-python = "^8.1.0"
aiohttp = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
